requirements.txt
environs==4.2.0
factory-boy==2.12.0
Faker==1.0.7
fire==0.1.3
google-auth==1.6.3
grpcio==1.22.0
grpcio-tools==1.22.0
kubernetes==10.0.1
MarkupSafe==1.1.1
marshmallow==2.19.5
pymysql==0.9.3
protobuf==3.9.1
py==1.8.0
pyasn1==0.4.7
pyasn1-modules==0.2.6
pylint==2.3.1
#pymilvus-test==0.2.15
pymilvus==0.2.0
pyparsing==2.4.0
pytest==4.6.3
pytest-level==0.1.1
pytest-print==0.1.2
pytest-repeat==0.8.0
pytest-timeout==1.3.3
python-dateutil==2.8.0
python-dotenv==0.10.3
pytz==2019.1
requests==2.22.0
requests-oauthlib==1.2.0
rsa==4.0
six==1.12.0
SQLAlchemy==1.3.5
urllib3==1.25.3
jaeger-client>=3.4.0
grpcio-opentracing>=1.0
